dev_requirements.txt
Flask-Migrate
Flask-Script
nose
rednose
Flask-Testing
colour-runner
coverage
